cleaned up formatting, conformer sampling improvements, and symmetry number started
